Move detectors to autonomous.vision package
