+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/test/opencv/WebcamStreamTest.java
@@ -5,7 +5,6 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.common.sensors.vision.WebcamStream;
-import org.firstinspires.ftc.teamcode.autonomous.util.opencv.ShapeGoldDetector;
 import org.opencv.android.OpenCVLoader;
 
 @Autonomous(name="Webcam Stream Test")
